+++ tools/onnx-graphsurgeon/tests/test_exporters.py
@@ -4,7 +4,7 @@ from onnx_graphsurgeon.logger.logger import G_LOGGER
 
 from onnx_models import identity_model, lstm_model, scan_model, dim_param_model
 
-from onnx_graphsurgeon.ir.tensor import Tensor, ConstantTensor, VariableTensor
+from onnx_graphsurgeon.ir.tensor import Tensor, Constant, Variable
 from onnx_graphsurgeon.ir.graph import Graph
 from onnx_graphsurgeon.ir.node import Node
 
@@ -15,20 +15,12 @@ import pytest
 import onnx
 
 class TestOnnxExporter(object):
-    def test_export_base_tensor(self):
-        name = "test_tensor"
-
-        tensor = Tensor(name)
-        onnx_tensor = OnnxExporter.export_value_info_proto(tensor)
-        assert onnx_tensor.name == name
-
-
     def test_export_constant_tensor_to_tensor_proto(self):
         name = "constant_tensor"
         shape = (3, 224, 224)
         values = np.random.random_sample(size=shape).astype(np.float32)
 
-        tensor = ConstantTensor(name=name, values=values)
+        tensor = Constant(name=name, values=values)
         onnx_tensor = OnnxExporter.export_tensor_proto(tensor)
         assert onnx_tensor.name == name
         assert np.all(onnx.numpy_helper.to_array(onnx_tensor) == values)
@@ -41,7 +33,7 @@ class TestOnnxExporter(object):
         shape = (3, 224, 224)
         values = np.random.random_sample(size=shape).astype(np.float32)
 
-        tensor = ConstantTensor(name=name, values=values)
+        tensor = Constant(name=name, values=values)
         onnx_tensor = OnnxExporter.export_value_info_proto(tensor)
         assert onnx_tensor.name == name
         assert onnx_tensor.type.tensor_type.elem_type == onnx.TensorProto.FLOAT
@@ -57,7 +49,7 @@ class TestOnnxExporter(object):
         shape = (3, 224, 224)
         dtype = np.float32
 
-        tensor = VariableTensor(dtype=dtype, shape=shape, name=name)
+        tensor = Variable(dtype=dtype, shape=shape, name=name)
         onnx_tensor = OnnxExporter.export_value_info_proto(tensor)
         assert onnx_tensor.name == name
         assert onnx_tensor.type.tensor_type.elem_type == onnx.TensorProto.FLOAT
@@ -72,13 +64,13 @@ class TestOnnxExporter(object):
     def test_export_node(self):
         name = "TestNode"
         op = "Test"
-        inputs = [Tensor(name="input")]
-        outputs = [Tensor(name="output")]
+        inputs = [Variable(name="input")]
+        outputs = [Variable(name="output")]
         attrs = OrderedDict()
         attrs["float_attr"] = 4.0
         attrs["int_attr"] = 10
         attrs["str_attr"] = "constant"
-        attrs["tensor_attr"] = ConstantTensor("testTensor", np.ones(shape=(1, 2, 3, 4), dtype=np.float32))
+        attrs["tensor_attr"] = Constant("testTensor", np.ones(shape=(1, 2, 3, 4), dtype=np.float32))
         attrs["floats_attr"] = [1.0, 2.0, 3.0, 4.0]
         attrs["ints_attr"] = [4, 3, 2, 1]
         attrs["strings_attr"] = ["constant", "and", "variable"]